+++ WordPress/src/main/java/org/wordpress/android/ui/blaze/BlazeParentActivity.kt
@@ -7,7 +7,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import org.wordpress.android.R
 import org.wordpress.android.ui.blaze.ui.blazeoverlay.BlazeOverlayFragment
 import org.wordpress.android.ui.blaze.ui.blazeoverlay.BlazeViewModel
-import org.wordpress.android.ui.blaze.ui.blazewebview.BlazeWebViewFragment
+import org.wordpress.android.ui.blaze.ui.blazepromote.BlazeWebViewFragment
 import org.wordpress.android.util.extensions.getParcelableExtraCompat
 import org.wordpress.android.util.extensions.getSerializableExtraCompat
