+++ WordPress/src/main/java/org/wordpress/android/ui/blaze/ui/blazepromote/BlazeWebViewFragment.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.blaze.ui.blazewebview
+package org.wordpress.android.ui.blaze.ui.blazepromote
 
 import android.annotation.SuppressLint
 import android.content.Intent
